Migrate depth test from legacy to pytest
